added:
A transformation to rename any Lustre keywords.
A utility to place Lustre specific data for use during translation
A class to translate declarations (not finished, but stubbed in)
A validation pass to highlight Lustre keywords in use (not sure if this
will remain, but it's stubbed in for now)

modifications:

minor changes to support translation in existing files.