checking in the new version of the translation.
this will better support the more advanced features. instead of
translating from Spear directly to Lustre the translation now collects
the concepts into an intermediate format (SpearNode, SpearProgram) that
has all the relevant elements. This makes translation a lot easier.

Also added the NameManager class which makes it very easy to obtain
unique names (and the code is a lot cleaner) with its
getUniqueNameAndRegister and lookup methods.